gateway/pyproject.toml
[project]
name = "gateway"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Fahad Ahmed", email = "fahad.ahmed95@live.com" }
]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.6",
    "numpy>=2.2.0",
    "pillow>=11.0.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.3",
    "uvicorn>=0.34.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
